Add /api/v1/import/ and Correct final item "Deploys a Function with the provided code."
Change 1) Add /api/v1/import/ - this is only applicable to 6.5 and above

The /api/v1/import will be used during an upgrade to import Eventing Functions created in 5.5.X or 6.0,X (e.g. pre 6.5.X) as langauge_compatibility field will be is not be present in any export or backup.   Upon import into a 6.5.X (or later) system via the REST endpoint  /api/v1/import the langauge_compatibility set to 6.0.0, thus this REST endpoint is important for seamless upgrades.

This is unlike the /api/v1/functions that are for used creating new Eventing Functions which if "langauge_compatibility" field is missing will set the value to the highest version supported by the server.

Currently "langauge_compatibility" has only two legal values 6.0.0 and 6.5.0

For details refer to https://issues.couchbase.com/browse/MB-37040 also "langauge_compatibility" causes Eventing to behave differently see https://issues.couchbase.com/browse/MB-32244 

Change 2) Correct last item "Deploys a Function with the provided code." - this may be applicable prior to 6.5

Although if we are in a paused state I believe we should be able to update code or "appcode" but this does not seem in any state to work as such I altered the word 'code' to 'settings'.  Of course the user can use /api/v1/functions/[function_name]/settings or /api/v1/import/ to update the "appcode" so this lack of functionality is not critical.

also updated to give an example of changing a setting via this REST endpoint and stress that changing a setting or code should only be done undeployed Functions.  In the example given if the payload was applied to a deployed Eventing Function as '{"deployment_status":true,"processing_status":true,"worker_count":9}' the value would change BUT no new workers would be started on the system.