DOC-4989-6.0 Update Required Ports (#777)
This is a backport of several changes made in the 6.5 release.

* Reformat and update required ports

This change targets the following issues:
* DOC-3972
* DOC-4397
* DOC-4511
* DOC-4989
* DOC-5267
* DOC-5795

There are still several changes that remain. This commit is one of several that will resolve this issue.

* Add note about FTS encrypted port usage

This change targets the following issues:
* DOC-3972
* DOC-5267
* DOC-5299

* Update release notes to reflect FTS changes

This change targets the following issues:
* DOC-3972
* DOC-5267
* DOC-5299

* Fix broken footnotes

This change targets the following issues:
* DOC-4989

Used passthrough macro with normoal substitutions (pass:n[]) to get xrefs to stop breaking. This should generate properly in Antora 3 as well.

* Added new fts footnote to comm path table