Reference the controlling Eventing 6.5 spec (#1257)
Disclaimer added as per Siri's request that the Couchbase 6.5 Eventing Specification is the controlling document, such that the document set is not considered a contract of behavior and functionality.  Engineering stands behind the specification on GitHub not the expanded documentation on docs.couchbase.com

Adding a note to this page seemed like the most logical place.