Fixing link to node SDK
The link is broken on the site. I *think* it's just a matter of an extra colon?

Also, why does this link just to node? Would it be okay to change this to something like: 

Execute N1QL queries programmatically using the official Couchbase SDKs: <Java> <.NET> <Node> <Python> <PHP> <Scala>

So the reader can choose the path they want with a single click instead of being dropped to node and then having to figure out how to switch to a different SDK?