Update createindex.adoc
CREATE INDEX is logically one of the first pages a new developer will see. With no intro to Couchbase, that user would be confused by this page. Too much complicated stuff right at the top. Too many warnings and caveats. This developer needs the basics first; then more details. I removed a LOT of content from the top of the page and made the language more consistent and succinct.