Update cluster-manager.adoc
In the documentation to this time, vBuckets were named as active and replica vBuckets; suddenly, in this page the active vBucket was referred to as "Master" vBucket thereby reducing readability and adding cognitive load to the reader. Hence changed the Master vBuckets to be called as active vBuckets to be consistent across all documentations.