DOC-9619: Updated information on Eventing and n2n for 7.1 (#2415)
* DOC-9619: Updated information on Eventing and n2n for 7.1

* Add CLI examples for Eventing

* clarity

An Eventing function doesn't run on just a node it is a cluster wide sharded function (on eventing nodes).  If you pause a function you only need to talk to one node and it pauses the function on all nodes running the eventing service.
Only resume the functions when encryption has been changed for all nodes (not one at a time)

Co-authored-by: Tony Hillman <tonyhillman@Tonys-MacBook-Pro-2.local>
Co-authored-by: Jon Strabala <54073900+jon-strabala@users.noreply.github.com>