Update list-cluster-nodes.adoc
In the List Nodes with the REST API section, updated REST API command to output both the otpNode and hostname.  This is because you need the otpNode value to be able to get detailed information about the node.  Then, added an example of how to use the `/nodes/<otpNode>` URI to display detailed information about the node.