Update rest-initialize-node.adoc to fix wrong field/key name in syntax and example (#3702)
For the REST API shown, both the syntax and the example showed the wrong field/key name for the data service path -- it is just  "path" in this specific REST API `/nodes/self/controller/settings` .   The other names are correct -- "index_path", "eventing_path", and "cbas_path".  Note that in the `/clusterInit` REST API, the field/key names are different.