[DOC-12942]: Create a release note for Couchbase Server 7.6.6
Tidy up.

Fixed the release date.

Adding entries to the release note.

Document fixes related to replica creation with alternateID handling and the use of `ALTER INDEX` with `replica_count` when shard affinity is enabled. These updates aim to clarify changes and prevent issues during rebalancing operations.

Updated the 7.6.6 release notes with newly resolved issues for the Storage Engine and Query Service. Highlights include a fix for high CPU usage in plasma items and a resolution for prepared statement malfunctions during rolling upgrades.

Remove redundant heading for version 7.6.4 features

The heading for "New Features and Enhancements in 7.6.4" was unnecessary as the features are already included via partials. This simplifies the structure and avoids duplication in the documentation.

This commit introduces the release notes for Couchbase Server 7.6.6, highlighting fixes for known issues and providing details on user-defined functions, sequences, and statistics data behavior after a bucket flush. It also includes workarounds to mitigate these issues.