DOC-12489 magma default storage engine (#3813)
This PR covers the impact of MB-62777 making Magma with 128 vBuckets the default storage backend.

It also tackles some of the work for DOC-12778 Data Settings guidance for reader/writer threads change to 'disk i/o optimized' needs to be revised because it was in the the same area of the doc being updated anyhow. The other areas of that ticket are being handled in the DOC-12485 prevent bucket from running out of space PR.